requirements.txt
django==4.1.4
django-restframework==3.14.0
